python/frameworks/pydantic-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "traceai-pydantic-ai"
version = "0.1.0"
description = "OpenTelemetry instrumentation for Pydantic AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "FutureAGI", email = "support@futureagi.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
keywords = ["pydantic-ai", "opentelemetry", "tracing", "observability", "llm", "agents"]

dependencies = [
    "opentelemetry-api>=1.0.0",
    "opentelemetry-sdk>=1.0.0",
    "opentelemetry-instrumentation>=0.40b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pydantic-ai>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/future-agi/traceai"
Documentation = "https://docs.futureagi.com/traceai"
Repository = "https://github.com/future-agi/traceai"

[tool.hatch.build.targets.wheel]
packages = ["traceai_pydantic_ai"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require pydantic-ai and API keys)",
]
